add testing to test Model & View

+++ LittleLemon/Restaurant/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Menu(models.Model):
     title = models.CharField(max_length = 255, db_index = True)
     price = models.DecimalField(max_digits = 6, decimal_places = 2, db_index = True)
-    inventory = models.IntegerField(db_index = True)
+    inventory = models.SmallIntegerField(db_index = True)
 
     def __str__(self) -> str:
         return f"{self.title} : {str(self.price)}"
